requirements.txt
discord
bs4
beautifulsoup4
Pillow==9.5.0
aiohttp
requests